Refactored examples and teaching examples
