+++ .gitignore
@@ -27,3 +27,5 @@ hs_err_pid*
 .settings
 .project
 .vscode
+
+target/
