+++ de.hamstersimulator.objectsfirst.teaching/pom.xml
@@ -1,0 +1,34 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <artifactId>teaching</artifactId>
+    <name>Hamstersimulator Teaching Module</name>
+    <description>
+        Classes used in the PSE lecture to teach programming. The classes here are designed to introduce
+        language concepts step by step.
+    </description>
+
+    <parent>
+        <groupId>de.hamstersimulator.objectsfirst</groupId>
+        <artifactId>parent</artifactId>
+        <version>5.0.1-SNAPSHOT</version>
+        <relativePath>../de.hamstersimulator.objectsfirst.parent</relativePath>
+    </parent>
+
+    <dependencies>
+        <dependency>
+            <groupId>
+                de.hamstersimulator.objectsfirst
+            </groupId>
+            <artifactId>main</artifactId>
+            <version>5.0.1-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>
+                de.hamstersimulator.objectsfirst
+            </groupId>
+            <artifactId>inspector</artifactId>
+            <version>5.0.1-SNAPSHOT</version>
+        </dependency>
+    </dependencies>
+</project>
