+++ de.hamstersimulator.objectsfirst.teaching/src/main/java/de/hamstersimulator/objectsfirst/teaching/advanced/GrainHamsters.java
@@ -1,4 +1,4 @@
-package de.hamstersimulator.objectsfirst.examples;
+package de.hamstersimulator.objectsfirst.teaching.advanced;
 
 import java.util.Collections;
 import java.util.HashSet;
@@ -7,12 +7,6 @@ import java.util.Set;
 import de.hamstersimulator.objectsfirst.external.model.Hamster;
 
 public class GrainHamsters {
-
-    public static void main(String[] args) {
-
-
-    }
-
     /**
      * ...
      * @param hamsters
