+++ de.hamstersimulator.objectsfirst.teaching/src/main/java/de/hamstersimulator/objectsfirst/teaching/lecture02/Example01.java
@@ -1,6 +1,6 @@
-package de.hamstersimulator.objectsfirst.examples;
+package de.hamstersimulator.objectsfirst.teaching.lecture02;
 
-class Example1 extends SimpleHamsterGame {
+class Example01 extends SimpleHamsterGame {
 
     @Override
     protected void run() {
