+++ de.hamstersimulator.objectsfirst.teaching/src/main/java/de/hamstersimulator/objectsfirst/teaching/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Hamstersimulator example classes
+ * @author Steffen Becker
+ *
+ */
+package de.hamstersimulator.objectsfirst.teaching;
