+++ de.hamstersimulator.objectsfirst.teaching/src/main/java/module-info.java
@@ -1,0 +1,4 @@
+module de.hamstersimulator.objectsfirst.teaching {
+    requires de.hamstersimulator.objectsfirst.main;
+    requires de.hamstersimulator.objectsfirst.ui;
+}
