+++ de.hamstersimulator.objectsfirst.teaching/src/main/resources/territory.ter
